Fix numerical stability issues in AMP mixed-precision training
- Replace entmax_bisect with internal entmax_beta implementation
  entmax_bisect internally forces float32 upcast which breaks AMP
  memory optimization. entmax_beta (sparsemax-based) supports
  alpha=1.5 without this issue and is compatible with AMP
- Add contiguous() after permute in hilbert_entmax
  Ensures memory is contiguous before reshape, avoids implicit copy
- swiglu.py: Remove forward内 clear_diagnostics()
  Avoids Python dict modification causing torch.compile graph break
  Cache cleared by ffn_output property after access instead
- depth_utils.py: Match dtype between regions and S_total tensor
  Prevents dtype mismatch warnings in log() operation

Improves AMP compatibility and reduces NaN risk

+++ src/vit_pytorch/core/depth_utils.py
@@ -476,7 +476,7 @@ def compute_normalized_area(
 
     # 面积归一化: f = log(s+1) / log(S+1)
     log_s_plus_1 = torch.log(s_patch + epsilon)
-    log_S_plus_1 = torch.log(torch.tensor(S_total + 1, device=regions.device, dtype=torch.float32))
+    log_S_plus_1 = torch.log(torch.tensor(S_total + 1, device=regions.device, dtype=regions.dtype))
     f_area = log_s_plus_1 / log_S_plus_1
 
     # 确保在 [0, 1] 范围内
